REFACTOR: move TeleDecl.StructDecl to ClassDecl.StructDecl
