+++ base/src/main/java/org/aya/core/def/FieldDef.java
@@ -4,6 +4,7 @@ package org.aya.core.def;
 
 import kala.collection.immutable.ImmutableSeq;
 import kala.control.Option;
+import org.aya.concrete.stmt.ClassDecl;
 import org.aya.concrete.stmt.TeleDecl;
 import org.aya.core.Matching;
 import org.aya.core.term.Term;
@@ -11,14 +12,14 @@ import org.aya.ref.DefVar;
 import org.jetbrains.annotations.NotNull;
 
 public final class FieldDef extends SubLevelDef {
-  public final @NotNull DefVar<StructDef, TeleDecl.StructDecl> structRef;
+  public final @NotNull DefVar<StructDef, ClassDecl.StructDecl> structRef;
   public final @NotNull DefVar<FieldDef, TeleDecl.StructField> ref;
   public final @NotNull Option<Term> body;
 
   public FieldDef(
-    @NotNull DefVar<StructDef, TeleDecl.StructDecl> structRef, @NotNull DefVar<FieldDef, TeleDecl.StructField> ref,
-    @NotNull ImmutableSeq<Term.Param> ownerTele, @NotNull ImmutableSeq<Term.Param> selfTele,
-    @NotNull Term result, @NotNull ImmutableSeq<Matching> clauses, @NotNull Option<Term> body, boolean coerce
+          @NotNull DefVar<StructDef, ClassDecl.StructDecl> structRef, @NotNull DefVar<FieldDef, TeleDecl.StructField> ref,
+          @NotNull ImmutableSeq<Term.Param> ownerTele, @NotNull ImmutableSeq<Term.Param> selfTele,
+          @NotNull Term result, @NotNull ImmutableSeq<Matching> clauses, @NotNull Option<Term> body, boolean coerce
   ) {
     super(ownerTele, selfTele, result, clauses, coerce);
     ref.core = this;
